# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PEPL API",
    "version": "1.0.0",
    "description": "Multi-tenant HR and employee-operations API.\n\n## Conventions\n\n- **Tenancy is implicit.** The company is resolved from the session token. There is no\n  tenant id in any path, header or body, and supplying one would have no effect.\n- **404, not 403, for scope.** A record belonging to another employee outside the caller's\n  data scope returns 404. Confirming that a record exists is itself a disclosure.\n- **Stable error codes.** Every error carries `error.code` — switch on that, not on the\n  status or the message. `error.requestId` correlates with the server log.\n- **Reasons.** Any change that moves money or rewrites history requires a `reason`, and\n  the API refuses the request without one.\n\n## Building a UI against this\n\nCall `GET /api/v1/me` first. It returns the caller's `permissions`, data `scope` and the\nset of enabled `modules`. Render navigation from that rather than from the role name —\na tenant can define custom roles, so role names are not a reliable switch.\n\nA module the tenant has disabled, or has not purchased, returns 403 `MODULE_NOT_AVAILABLE`.\nHide it rather than showing a disabled teaser."
  },
  "servers": [
    {
      "url": "http://localhost:4010",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/leave/balances": {
      "get": {
        "summary": "Leave balances, explained",
        "tags": [
          "leave"
        ],
        "operationId": "get_api_v1_leave_balances",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, or the module is not enabled for this company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found — also returned for a record outside the caller's data scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (a frozen period, a locked run)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed, or a domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Requires the `leave.read` permission."
      }
    },
    "/api/v1/leave/balances/{leaveTypeId}/ledger": {
      "get": {
        "summary": "Every movement behind a balance",
        "tags": [
          "leave"
        ],
        "operationId": "get_api_v1_leave_balances__leaveTypeId__ledger",
        "parameters": [
          {
            "name": "leaveTypeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, or the module is not enabled for this company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found — also returned for a record outside the caller's data scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (a frozen period, a locked run)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed, or a domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Requires the `leave.read` permission."
      }
    },
    "/api/v1/leave/requests": {
      "post": {
        "summary": "Apply for leave",
        "tags": [
          "leave"
        ],
        "operationId": "post_api_v1_leave_requests",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, or the module is not enabled for this company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found — also returned for a record outside the caller's data scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (a frozen period, a locked run)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed, or a domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Requires the `leave.apply` permission.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "leaveTypeId": "…",
                "startDate": "2026-09-14",
                "endDate": "2026-09-16",
                "totalDays": 3,
                "reason": "family event"
              }
            }
          }
        }
      }
    },
    "/api/v1/leave/requests/{id}/cancel": {
      "post": {
        "summary": "Cancel leave — writes a reversal, never deletes the consumption",
        "tags": [
          "leave"
        ],
        "operationId": "post_api_v1_leave_requests__id__cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, or the module is not enabled for this company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found — also returned for a record outside the caller's data scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (a frozen period, a locked run)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed, or a domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Requires the `leave.apply` permission."
      }
    },
    "/api/v1/leave/balances/adjust": {
      "post": {
        "summary": "Adjust a balance directly (HR only, reason required)",
        "tags": [
          "leave"
        ],
        "operationId": "post_api_v1_leave_balances_adjust",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, or the module is not enabled for this company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found — also returned for a record outside the caller's data scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (a frozen period, a locked run)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed, or a domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Requires the `leave.balance.adjust` permission.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "employeeId": "…",
                "leaveTypeId": "…",
                "days": 2,
                "reason": "opening balance correction"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The token from POST /api/v1/auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine code. Switch on this."
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        },
        "example": {
          "error": {
            "code": "CONFIG_EFFECTIVE_DATE_REQUIRED",
            "message": "\"payroll.lop_basis\" affects payroll and must carry an effective date, so a locked run stays reproducible",
            "requestId": "…"
          }
        }
      }
    }
  },
...
}